20240731/requirements.txt
psycopg2-binary
python-dotenv   
flask
gunicorn
Flask-WTF